Move ModifierPanel to UI package
